requirements.txt
boto3
pypdf
streamlit
faiss-cpu
watchdog
llama-index
langchain-community

